backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pgvector==0.2.4

# AI/ML
google-generativeai==0.3.1
sentence-transformers==2.2.2
numpy==1.26.2
torch==2.1.1

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# Web Scraping (for job data)
httpx==0.25.2
beautifulsoup4==4.12.2
playwright==1.40.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
tenacity==8.2.3

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
